deck state module
